requirements.txt
django==2.1.4
django-mptt==0.9.1
django-rest-swagger==2.2.0